Move go/sig/internal/disp to go/lib
