+++ go/lib/sigdisp/disp.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package disp
+package sigdisp
 
 import (
 	"fmt"
