+++ go/sig/BUILD.bazel
@@ -13,10 +13,10 @@ go_library(
         "//go/lib/log:go_default_library",
         "//go/lib/prom:go_default_library",
         "//go/lib/serrors:go_default_library",
+        "//go/lib/sigdisp:go_default_library",
         "//go/sig/config:go_default_library",
         "//go/sig/egress:go_default_library",
         "//go/sig/internal/base:go_default_library",
-        "//go/sig/internal/disp:go_default_library",
         "//go/sig/internal/ingress:go_default_library",
         "//go/sig/internal/metrics:go_default_library",
         "//go/sig/internal/sigcmn:go_default_library",
