+++ go/sig/egress/session/BUILD.bazel
@@ -18,12 +18,12 @@ go_library(
         "//go/lib/pathmgr:go_default_library",
         "//go/lib/pktdisp:go_default_library",
         "//go/lib/ringbuf:go_default_library",
+        "//go/lib/sigdisp:go_default_library",
         "//go/lib/snet:go_default_library",
         "//go/lib/spath/spathmeta:go_default_library",
         "//go/sig/egress/iface:go_default_library",
         "//go/sig/egress/siginfo:go_default_library",
         "//go/sig/egress/worker:go_default_library",
-        "//go/sig/internal/disp:go_default_library",
         "//go/sig/internal/metrics:go_default_library",
         "//go/sig/internal/sigcmn:go_default_library",
     ],
